(feat): use external templates in theme (#20)
* Update resource box button class

* Remove 'authored_by' from quote block

* Refactor file tag

* Use liquid templates in theme
