+++ lib/jekyll-plugin-platoniq-journal.rb
@@ -4,6 +4,8 @@ require "jekyll"
 require "kramdown"
 require "jekyll-plugin-platoniq-journal/base"
 require "jekyll-plugin-platoniq-journal/includes_file"
+require "jekyll-plugin-platoniq-journal/external_template_block"
+require "jekyll-plugin-platoniq-journal/external_template_tag"
 require "jekyll-plugin-platoniq-journal/gallery_tag"
 require "jekyll-plugin-platoniq-journal/call_to_action_block"
 require "jekyll-plugin-platoniq-journal/file_tag"
