+++ lib/jekyll-plugin-platoniq-journal/base.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-# require "byebug"
-
 module JekyllPluginPlatoniqJournal
   module Base
     def site
